unction that prints the binary representation of a number

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -9,7 +9,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned int i = 1;
+	unsigned long int i = 1;
 
 	for (; i < n; i <<= 1)
 	{
